Fixed to work with API changes, and new "retrieve mode" option.
Works with the latest API, which also included changes to the way passwords
were encrypted (credits to @IkaEren for reversing the new crypto code).

Retrieve mode uses the credential retrieval API (retrieve_user_*), which is
meant to get a new set of credentials for an existing user. The default
create mode API creates a new user each time it is called, which will now
lead to errors if you call it too many times.

Also some minor code cleanup.